# swagger-spec.json
{"openapi":"3.0.0","paths":{"/students":{"post":{"operationId":"StudentsController_createStudent","summary":"can access: sputnik, curator","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateStudentDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StudentDto"}}}}},"tags":["Students"],"security":[{"bearer":[]}]}},"/students/{uuid}/ban":{"patch":{"operationId":"StudentsController_banStudent","summary":"can access: sputnik, curator","parameters":[{"name":"uuid","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateResult"}}}}},"tags":["Students"],"security":[{"bearer":[]}]}},"/students/{uuid}/unban":{"patch":{"operationId":"StudentsController_unbanStudent","summary":"can access: sputnik, curator","parameters":[{"name":"uuid","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateResult"}}}}},"tags":["Students"],"security":[{"bearer":[]}]}},"/groups/{id}/students":{"get":{"operationId":"StudentsController_getStudentsByGroup","summary":"can access: all","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/StudentDto"}}}}}},"tags":["Students"],"security":[{"bearer":[]}]}},"/institutes/{id}/students":{"get":{"operationId":"StudentsController_getStudentsByInstitute","summary":"can access: all","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/StudentDto"}}}}}},"tags":["Students"],"security":[{"bearer":[]}]}},"/students/me":{"get":{"operationId":"StudentsController_getMe","summary":"can access: student","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StudentDto"}}}}},"tags":["Students"],"security":[{"bearer":[]}]}},"/sputniks":{"post":{"operationId":"SputniksController_createSputnik","summary":"can access: curator","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSputnikDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SputnikDto"}}}}},"tags":["Sputniks"],"security":[{"bearer":[]}]}},"/groups/{id}/sputniks":{"get":{"operationId":"SputniksController_getSputniksByGroup","summary":"can access: all","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SputnikDto"}}}}}},"tags":["Sputniks"],"security":[{"bearer":[]}]}},"/institutes/{id}/sputniks":{"get":{"operationId":"SputniksController_getSputniksByInstitute","summary":"can access: all","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SputnikDto"}}}}}},"tags":["Sputniks"],"security":[{"bearer":[]}]}},"/sputniks/me":{"get":{"operationId":"SputniksController_getMe","summary":"can access: sputnik","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SputnikDto"}}}}},"tags":["Sputniks"],"security":[{"bearer":[]}]}},"/curator":{"post":{"operationId":"CuratorsController_createCurator","summary":"can access: admin only","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCuratorDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CuratorDto"}}}}},"tags":["Curators"]}},"/institutes/{id}/curators":{"get":{"operationId":"CuratorsController_getCuratorsByInstitute","summary":"can access: all","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CuratorDto"}}}}}},"tags":["Curators"]}},"/curators/me":{"get":{"operationId":"CuratorsController_getMe","summary":"can access: curator","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CuratorDto"}}}}},"tags":["Curators"]}},"/groups":{"post":{"operationId":"GroupsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateGroupDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GroupDto"}}}}},"tags":["Groups"]}},"/institutes/{id}/groups":{"get":{"operationId":"GroupsController_getGroupsByInstitute","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GroupDto"}}}}}},"tags":["Groups"]}},"/institutes":{"get":{"operationId":"InstitutesController_getInstitutes","parameters":[],"responses":{"200":{"description":""}},"tags":["Institutes"]}},"/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthResponseDto"}}}}},"tags":["Auth"]}}},"info":{"title":"SUAI Achievements","description":"","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"CreateStudentDto":{"type":"object","properties":{"firstName":{"type":"string"},"lastName":{"type":"string"},"patronymic":{"type":"string"},"vkId":{"type":"string"},"instituteId":{"type":"number"},"groupId":{"type":"number"}},"required":["firstName","lastName","patronymic","vkId","instituteId","groupId"]},"InstituteDto":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string"},"shortName":{"type":"string"},"number":{"type":"number"}},"required":["id","name","shortName","number"]},"GroupDto":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string"}},"required":["id","name"]},"StudentDto":{"type":"object","properties":{"firstName":{"type":"string"},"lastName":{"type":"string"},"patronymic":{"type":"string"},"role":{"type":"string"},"isBanned":{"type":"boolean"},"institute":{"$ref":"#/components/schemas/InstituteDto"},"group":{"$ref":"#/components/schemas/GroupDto"},"balance":{"type":"number"}},"required":["firstName","lastName","patronymic","role","isBanned","institute","group","balance"]},"UpdateResult":{"type":"object","properties":{}},"CreateSputnikDto":{"type":"object","properties":{"firstName":{"type":"string"},"lastName":{"type":"string"},"patronymic":{"type":"string"},"vkId":{"type":"string"},"instituteId":{"type":"number"},"groupIds":{"example":[0,1],"type":"array","items":{"type":"array"}}},"required":["firstName","lastName","patronymic","vkId","instituteId","groupIds"]},"SputnikDto":{"type":"object","properties":{"firstName":{"type":"string"},"lastName":{"type":"string"},"patronymic":{"type":"string"},"role":{"type":"string"},"isBanned":{"type":"boolean"},"institute":{"$ref":"#/components/schemas/InstituteDto"},"sputnikGroups":{"type":"array","items":{"type":"array"}}},"required":["firstName","lastName","patronymic","role","isBanned","institute","sputnikGroups"]},"CreateCuratorDto":{"type":"object","properties":{"firstName":{"type":"string"},"lastName":{"type":"string"},"patronymic":{"type":"string"},"vkId":{"type":"string"},"instituteId":{"type":"number"}},"required":["firstName","lastName","patronymic","vkId","instituteId"]},"CuratorDto":{"type":"object","properties":{"firstName":{"type":"string"},"lastName":{"type":"string"},"patronymic":{"type":"string"},"role":{"type":"string"},"isBanned":{"type":"boolean"},"institute":{"$ref":"#/components/schemas/InstituteDto"}},"required":["firstName","lastName","patronymic","role","isBanned","institute"]},"CreateGroupDto":{"type":"object","properties":{"name":{"type":"string"},"instituteId":{"type":"number"}},"required":["name","instituteId"]},"AuthDto":{"type":"object","properties":{"vkId":{"type":"string"},"vkToken":{"type":"string"}},"required":["vkId","vkToken"]},"AuthResponseDto":{"type":"object","properties":{"accessToken":{"type":"string"},"role":{"type":"string"}},"required":["accessToken","role"]}}}}